change endpoint section
