+++ core/src/main/java/greencity/controller/ManagementOrderController.java
@@ -40,12 +40,14 @@ import greencity.dto.violation.UpdateViolationToUserDto;
 import greencity.dto.violation.ViolationDetailInfoDto;
 import greencity.dto.violation.ViolationsInfoDto;
 import greencity.entity.parameters.CustomTableView;
+import greencity.entity.user.User;
 import greencity.filters.CertificateFilterCriteria;
 import greencity.filters.CertificatePage;
 import greencity.filters.OrderPage;
 import greencity.filters.OrderSearchCriteria;
 import greencity.service.ubs.CertificateService;
 import greencity.service.ubs.CoordinateService;
+import greencity.service.ubs.UBSClientService;
 import greencity.service.ubs.UBSManagementService;
 import greencity.service.ubs.ViolationService;
 import greencity.service.ubs.manager.BigOrderTableServiceView;
@@ -87,6 +89,7 @@ import springfox.documentation.annotations.ApiIgnore;
 @RequiredArgsConstructor
 public class ManagementOrderController {
     private final UBSManagementService ubsManagementService;
+    private final UBSClientService ubsClientService;
     private final CertificateService certificateService;
     private final CoordinateService coordinateService;
     private final ViolationService violationService;
@@ -1012,5 +1015,29 @@ public class ManagementOrderController {
         @Valid @PathVariable("id") Long orderId) {
         return ResponseEntity.status(HttpStatus.OK)
             .body(ubsManagementService.getNotTakenOrderReason(orderId));
+    }
+
+    /**
+     * Controller updates info about order cancellation reason.
+     *
+     * @param id   {@link Long}.
+     * @param dto  {@link OrderCancellationReasonDto}
+     * @param uuid current {@link User}'s uuid.
+     * @return {@link HttpStatus} - http status.
+     */
+    @ApiOperation(value = "updates info about order cancellation reason ")
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = HttpStatuses.OK, response = OrderCancellationReasonDto.class),
+        @ApiResponse(code = 400, message = HttpStatuses.BAD_REQUEST),
+        @ApiResponse(code = 401, message = HttpStatuses.UNAUTHORIZED),
+        @ApiResponse(code = 403, message = HttpStatuses.FORBIDDEN),
+        @ApiResponse(code = 404, message = HttpStatuses.NOT_FOUND)
+    })
+    @PostMapping("/order/{id}/cancellation")
+    public ResponseEntity<OrderCancellationReasonDto> updateCancellationReason(
+        @RequestBody final OrderCancellationReasonDto dto,
+        @PathVariable("id") final Long id,
+        @ApiIgnore @CurrentUserUuid String uuid) {
+        return ResponseEntity.status(HttpStatus.OK).body(ubsClientService.updateOrderCancellationReason(id, dto, uuid));
     }
 }
